Run tiff tile source tests in the container as opposed to host
