+++ devops/ansible/large_image/molecule/tiff/tests/test_tiff_tile_source.py
@@ -1,10 +1,7 @@
 """Test the default libtiff based tile source."""
 
 import os
-import large_image
 import testinfra.utils.ansible_runner
-import urllib
-import pytest
 
 
 testinfra_hosts = testinfra.utils.ansible_runner.AnsibleRunner(
@@ -18,15 +15,9 @@ def test_pip_packages(host):
     assert 'libtiff' in packages
 
 
-@pytest.mark.parametrize("item, output", [
-    ('590346ff8d777f16d01e054c', '/tmp/Huron.Image2_JPEG2K.tif')
-])
-def test_tiff_tile_source(item, output):
-    """Check whether large_image can return a tile with tiff sources."""
-    test_url = 'https://data.kitware.com/api/v1/item/{}/download'.format(item)
-    urllib.urlretrieve(test_url, output)
-    image = large_image.getTileSource(output)
-    # Make sure it is the tiff tile source
-    assert isinstance(image, large_image.tilesource.TiffFileTileSource)
-    # Make sure we can get a tile without an exception
-    assert type(image.getTile(0, 0, 0)) == str
+def test_large_image_tiff_source(host):
+    """Run large image tiff tests."""
+    activate = "source /root/.virtualenvs/large_image/bin/activate"
+    run = "pytest /test_tiff_tile_source.py"
+    cmd = host.run("bash -c '{} && {}'".format(activate, run))
+    assert cmd.stderr == ''
